backend/requirements.txt
fastapi
uvicorn
snowflake-snowpark-python
python-dotenv
pandas
